+++ db/migrate/20220226201943_add_published_to_pages.rb
@@ -1,0 +1,6 @@
+class AddPublishedToPages < ActiveRecord::Migration[6.1]
+  def change
+    add_column :pages, :published, :boolean, null: false, default: false
+    add_index :pages, :published
+  end
+end